+++ src/main/java/google/com/ortona/hashcode/qualification_2017/model/SolutionContainer.java
@@ -18,7 +18,7 @@ public class SolutionContainer {
 		return score;
 	}
 
-	public void setScore(int score) {
+	public void setScore(double score) {
 		this.score = score;
 	}
